+++ lib/database/journal_db/config_flags.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 
 import 'package:lotti/database/database.dart';
 import 'package:lotti/utils/consts.dart';
-import 'package:lotti/utils/platform.dart';
 
 Future<void> initConfigFlags(JournalDb db) async {
   await db.insertFlagIfNotExists(
@@ -28,15 +27,8 @@ Future<void> initConfigFlags(JournalDb db) async {
   );
   await db.insertFlagIfNotExists(
     const ConfigFlag(
-      name: enableSyncInboxFlag,
-      description: 'Enable sync inbox? (requires restart)',
-      status: true,
-    ),
-  );
-  await db.insertFlagIfNotExists(
-    const ConfigFlag(
-      name: enableSyncOutboxFlag,
-      description: 'Enable sync outbox? (requires restart)',
+      name: enableSyncFlag,
+      description: 'Enable sync? (requires restart)',
       status: true,
     ),
   );
@@ -45,15 +37,6 @@ Future<void> initConfigFlags(JournalDb db) async {
       const ConfigFlag(
         name: enableNotificationsFlag,
         description: 'Enable desktop notifications?',
-        status: false,
-      ),
-    );
-  }
-  if (isDesktop) {
-    await db.insertFlagIfNotExists(
-      const ConfigFlag(
-        name: showThemeConfigFlag,
-        description: 'Show Theme Config UI?',
         status: false,
       ),
     );