+++ test/database/database_test.dart
@@ -5,17 +5,10 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/database/journal_db/config_flags.dart';
 import 'package:lotti/utils/consts.dart';
 
-final expectedActiveFlagNames = Platform.isMacOS
-    ? {
-        privateFlag,
-        enableSyncInboxFlag,
-        enableSyncOutboxFlag,
-      }
-    : {
-        privateFlag,
-        enableSyncInboxFlag,
-        enableSyncOutboxFlag,
-      };
+final expectedActiveFlagNames = {
+  privateFlag,
+  enableSyncFlag,
+};
 
 final expectedFlags = <ConfigFlag>{
   const ConfigFlag(
@@ -29,23 +22,13 @@ final expectedFlags = <ConfigFlag>{
     status: false,
   ),
   const ConfigFlag(
-    name: showThemeConfigFlag,
-    description: 'Show Theme Config UI?',
-    status: false,
-  ),
-  const ConfigFlag(
     name: allowInvalidCertFlag,
     description: 'Allow invalid certificate? (not recommended)',
     status: false,
   ),
   const ConfigFlag(
-    name: enableSyncInboxFlag,
-    description: 'Enable sync inbox? (requires restart)',
-    status: true,
-  ),
-  const ConfigFlag(
-    name: enableSyncOutboxFlag,
-    description: 'Enable sync outbox? (requires restart)',
+    name: enableSyncFlag,
+    description: 'Enable sync? (requires restart)',
     status: true,
   ),
 };